requirements.txt
openai
python-dotenv
argparse